requirements.txt
# Core
pydantic>=2.0.0
pydantic-settings>=2.0.0
sqlalchemy>=2.0.0
python-dotenv>=1.0.0
click>=8.0.0
pyyaml>=6.0
typer>=0.9.0

# API
fastapi-poe>=0.0.46
deepgram-sdk>=3.0.0

# Voice recording (опционально)
# sounddevice>=0.4.6
# wavio>=0.0.9

# Database
# psycopg2-binary>=2.9.0
# pymysql>=1.1.0

# Audio processing
pydub>=0.25.1
ffmpeg-python>=0.2.0
librosa>=0.10.0
numpy>=1.24.0
mutagen>=1.47.0

# Image processing
Pillow>=10.0.0
opencv-python>=4.8.0

# HTTP clients
requests>=2.31.0
httpx>=0.25.0
aiohttp>=3.9.0

# Browser automation
playwright>=1.40.0
selenium>=4.15.0

# Storage
boto3>=1.34.0
python-telegram-bot>=20.7

# Web framework (FastAPI)
fastapi>=0.109.0
uvicorn>=0.27.0
jinja2>=3.1.3
python-multipart>=0.0.6

# Utilities
tqdm>=4.66.0
rich>=13.7.0
matplotlib>=3.7.0

# Development
pytest>=7.4.0
black>=23.0.0
mypy>=1.7.0
